Update footer
Both the footer and the newly created modal component can now switch between a light and dark theme, depending on the page background. Some tweaks were made to the icons, as well as to the hover animations.

Although the footer is now displayed correctly on small displays, without the bottom padding, it ends up getting cut off on all pages except for the form confirmation page.

Closes: #77, #78, #26, #34